# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.20.0 🍺
 * Do not edit manually.
 * Api
 * SkyReserve Flight Booking Management System API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface Airline {
  id: number;
  name: string;
  country: string;
  /** @nullable */
  logoUrl?: string | null;
  createdAt?: string;
}

export interface AirlineInput {
  /** @minLength 1 */
  name: string;
  /** @minLength 1 */
  country: string;
  logoUrl?: string;
}

export interface AirlineUpdate {
  /** @minLength 1 */
  name?: string;
  /** @minLength 1 */
  country?: string;
  logoUrl?: string;
}

export interface AirlineStats {
  airlineId: number;
  fleetSize: number;
  activeFlights: number;
  totalBookings: number;
  totalRevenue: number;
}

export interface Flight {
  id: number;
  airlineId: number;
  airlineName: string;
  flightNumber?: string;
  departure: string;
  arrival: string;
  departureTime: string;
  arrivalTime: string;
  status: string;
  aircraft: string;
  totalSeats: number;
  availableSeats: number;
  basePrice: number;
  durationMinutes: number;
}

export interface FlightInput {
  airlineId: number;
  flightNumber?: string;
  /** @minLength 1 */
  departure: string;
  /** @minLength 1 */
  arrival: string;
  departureTime: string;
  arrivalTime: string;
  status?: string;
  /** @minLength 1 */
  aircraft: string;
  totalSeats: number;
  basePrice: number;
}

export interface FlightUpdate {
  airlineId?: number;
  flightNumber?: string;
  departure?: string;
  arrival?: string;
  departureTime?: string;
  arrivalTime?: string;
  status?: string;
  aircraft?: string;
  totalSeats?: number;
  basePrice?: number;
}

export interface Seat {
  id: number;
  flightId: number;
  seatNumber: string;
  cabinClass: string;
  status: string;
}

export interface Customer {
  id: number;
  name: string;
  email: string;
  phone: string;
  role: string;
  /** @nullable */
  avatarUrl?: string | null;
  darkModePref?: boolean;
  notifyEmail?: boolean;
  notifySms?: boolean;
  createdAt?: string;
}

export interface CustomerUpdate {
  /** @minLength 1 */
  name?: string;
  phone?: string;
  avatarUrl?: string;
  darkModePref?: boolean;
  notifyEmail?: boolean;
  notifySms?: boolean;
}

export interface CustomerAdminUpdate {
  name?: string;
  phone?: string;
  role?: string;
}

export interface Booking {
  id: number;
  customerId: number;
  customerName: string;
  flightId: number;
  bookingDate: string;
  seatNumber: string;
  cabinClass: string;
  status: string;
  price: number;
  bookingReference: string;
  passengerName?: string;
  passengerEmail?: string;
  passengerPhone?: string;
  flight: Flight;
}

export interface BookingInput {
  flightId: number;
  /** @minLength 1 */
  seatNumber: string;
  /** @minLength 1 */
  cabinClass: string;
  /** @minLength 1 */
  passengerName: string;
  /** @minLength 1 */
  passengerEmail: string;
  /** @minLength 1 */
  passengerPhone: string;
  paymentMethodId?: number;
}

export interface BookingUpdate {
  seatNumber?: string;
  status?: string;
  cabinClass?: string;
}

export interface Ticket {
  bookingId: number;
  bookingReference: string;
  passengerName: string;
  flight: Flight;
  seatNumber: string;
  cabinClass: string;
  qrPayload: string;
  status: string;
}

export interface SavedPassenger {
  id: number;
  customerId: number;
  fullName: string;
  email: string;
  phone: string;
  /** @nullable */
  passportNumber?: string | null;
  /** @nullable */
  dateOfBirth?: string | null;
}

export interface SavedPassengerInput {
  /** @minLength 1 */
  fullName: string;
  /** @minLength 1 */
  email: string;
  /** @minLength 1 */
  phone: string;
  passportNumber?: string;
  dateOfBirth?: string;
}

export interface PaymentMethod {
  id: number;
  customerId: number;
  cardBrand: string;
  last4: string;
  expiryMonth: number;
  expiryYear: number;
  cardholderName: string;
}

export interface PaymentMethodInput {
  /** @minLength 1 */
  cardBrand: string;
  /** @minLength 4 */
  cardNumber: string;
  expiryMonth: number;
  expiryYear: number;
  /** @minLength 1 */
  cardholderName: string;
}

export interface WishlistItem {
  id: number;
  customerId: number;
  flight: Flight;
}

export interface WishlistInput {
  flightId: number;
}

export interface Notification {
  id: number;
  customerId: number;
  title: string;
  message: string;
  isRead: boolean;
  createdAt: string;
}

export interface DashboardSummary {
  totalCustomers: number;
  totalFlights: number;
  totalBookings: number;
  totalRevenue: number;
  upcomingFlights: number;
  delayedFlights: number;
  cancelledFlights: number;
  confirmedBookings: number;
}

export interface RevenuePoint {
  period: string;
  revenue: number;
  bookings: number;
}

export interface OccupancyPoint {
  flightId: number;
  departure: string;
  arrival: string;
  totalSeats: number;
  bookedSeats: number;
  occupancyRate: number;
}

export interface TopCustomer {
  customerId: number;
  customerName: string;
  totalBookings: number;
  totalSpent: number;
}

export interface TopFlight {
  flightId: number;
  departure: string;
  arrival: string;
  totalBookings: number;
}

export interface AirlineComparison {
  airlineId: number;
  airlineName: string;
  totalFlights: number;
  totalBookings: number;
  totalRevenue: number;
  onTimeRate: number;
}

export interface ActivityEvent {
  id: number;
  type: string;
  message: string;
  createdAt: string;
}

export interface TriggerLog {
  id: number;
  triggerName: string;
  message: string;
  createdAt: string;
}

export interface ProcedureDefinition {
  name: string;
  label: string;
  description: string;
  params: string[];
}

export interface ProcedureExecutionInput {
  customerId?: number;
  flightId?: number;
  bookingDate?: string;
  seatNumber?: string;
  cabinClass?: string;
  status?: string;
  bookingId?: number;
}

export interface ProcedureExecutionResult {
  id: number;
  procedureName: string;
  success: boolean;
  output: string;
  createdAt: string;
}

export interface ConfirmedBookingRow {
  customerId: number;
  customerName: string;
  email: string;
  bookingId: number;
  departure: string;
  arrival: string;
  seatNumber: string;
  bookingDate: string;
}

export interface OccupancySummaryRow {
  flightId: number;
  airlineName: string;
  departure: string;
  arrival: string;
  confirmedBookings: number;
}

export interface QueryDefinition {
  id: string;
  label: string;
  description: string;
  category: string;
  sql: string;
}

export interface QueryResult {
  id: string;
  columns: string[];
  rows: unknown[];
  rowCount: number;
  executedAt: string;
}

export type ListFlightsParams = {
origin?: string;
destination?: string;
date?: string;
cabinClass?: string;
status?: string;
airlineId?: number;
minSeats?: number;
};

export type ListCustomersParams = {
search?: string;
};

export type ListBookingsParams = {
status?: string;
search?: string;
all?: boolean;
};