# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: SkyReserve Flight Booking Management System API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /flights:
    get:
      operationId: listFlights
      tags: [flights]
      parameters:
        - { name: origin, in: query, schema: { type: string } }
        - { name: destination, in: query, schema: { type: string } }
        - { name: date, in: query, schema: { type: string } }
        - { name: cabinClass, in: query, schema: { type: string } }
        - { name: status, in: query, schema: { type: string } }
        - { name: airlineId, in: query, schema: { type: integer } }
        - { name: minSeats, in: query, schema: { type: integer } }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/Flight" }
    post:
      operationId: createFlight
      tags: [flights]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/FlightInput" }
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Flight" }
  /flights/{id}:
    get:
      operationId: getFlight
      tags: [flights]
      parameters:
        - { name: id, in: path, required: true, schema: { type: integer } }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Flight" }
        "404": { description: Not found }
    patch:
      operationId: updateFlight
      tags: [flights]
      parameters:
        - { name: id, in: path, required: true, schema: { type: integer } }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/FlightUpdate" }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Flight" }
    delete:
      operationId: deleteFlight
      tags: [flights]
      parameters:
        - { name: id, in: path, required: true, schema: { type: integer } }
      responses:
        "204": { description: Deleted }
  /flights/{id}/seats:
    get:
      operationId: getFlightSeatMap
      tags: [flights]
      parameters:
        - { name: id, in: path, required: true, schema: { type: integer } }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/Seat" }
...
  /notifications/{id}/read:
    patch:
      operationId: markNotificationRead
      tags: [notifications]
      parameters:
        - { name: id, in: path, required: true, schema: { type: integer } }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Notification" }
...
  /db/procedures/{name}/execute:
    post:
      operationId: executeProcedure
      tags: [db]
      parameters:
        - { name: name, in: path, required: true, schema: { type: string } }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/ProcedureExecutionInput" }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ProcedureExecutionResult" }
...
  /db/queries/{id}/execute:
    post:
      operationId: executeQuery
      tags: [db]
      parameters:
        - { name: id, in: path, required: true, schema: { type: string } }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/QueryResult" }
components:
  schemas:
...
    Flight:
      type: object
      required: [id, airlineId, airlineName, departure, arrival, departureTime, arrivalTime, status, aircraft, totalSeats, availableSeats, basePrice, durationMinutes]
      properties:
        id: { type: integer }
        airlineId: { type: integer }
        airlineName: { type: string }
        flightNumber: { type: string }
        departure: { type: string }
        arrival: { type: string }
        departureTime: { type: string }
        arrivalTime: { type: string }
        status: { type: string }
        aircraft: { type: string }
        totalSeats: { type: integer }
        availableSeats: { type: integer }
        basePrice: { type: number }
        durationMinutes: { type: integer }
    FlightInput:
      type: object
      required: [airlineId, departure, arrival, departureTime, arrivalTime, aircraft, totalSeats, basePrice]
      properties:
        airlineId: { type: integer }
        flightNumber: { type: string }
        departure: { type: string, minLength: 1 }
        arrival: { type: string, minLength: 1 }
        departureTime: { type: string }
        arrivalTime: { type: string }
        status: { type: string }
        aircraft: { type: string, minLength: 1 }
        totalSeats: { type: integer }
        basePrice: { type: number }
    FlightUpdate:
      type: object
      properties:
        airlineId: { type: integer }
        flightNumber: { type: string }
        departure: { type: string }
        arrival: { type: string }
        departureTime: { type: string }
        arrivalTime: { type: string }
        status: { type: string }
        aircraft: { type: string }
        totalSeats: { type: integer }
        basePrice: { type: number }
    Seat:
      type: object
      required: [id, flightId, seatNumber, cabinClass, status]
      properties:
        id: { type: integer }
        flightId: { type: integer }
        seatNumber: { type: string }
        cabinClass: { type: string }
        status: { type: string }
...
    Notification:
      type: object
      required: [id, customerId, title, message, isRead, createdAt]
      properties:
        id: { type: integer }
        customerId: { type: integer }
        title: { type: string }
        message: { type: string }
        isRead: { type: boolean }
        createdAt: { type: string }
...
    ProcedureExecutionInput:
      type: object
      properties:
        customerId: { type: integer }
        flightId: { type: integer }
        bookingDate: { type: string }
        seatNumber: { type: string }
        cabinClass: { type: string }
        status: { type: string }
        bookingId: { type: integer }
    ProcedureExecutionResult:
      type: object
      required: [id, procedureName, success, output, createdAt]
      properties:
        id: { type: integer }
        procedureName: { type: string }
        success: { type: boolean }
        output: { type: string }
        createdAt: { type: string }
...
    QueryResult:
      type: object
      required: [id, columns, rows, rowCount, executedAt]
      properties:
        id: { type: string }
        columns:
          type: array
          items: { type: string }
        rows:
          type: array
          items: {}
        rowCount: { type: integer }
        executedAt: { type: string }