# lib/api-zod/src/generated/types/seat.ts
/**
 * Generated by orval v8.20.0 🍺
 * Do not edit manually.
 * Api
 * SkyReserve Flight Booking Management System API
 * OpenAPI spec version: 0.1.0
 */

export interface Seat {
  id: number;
  flightId: number;
  seatNumber: string;
  cabinClass: string;
  status: string;
}